alpha_loan_project/apps/communications/email/gmail_client.py
"""Gmail Client - Email API integration"""

import os
import base64
from google.oauth2.service_account import Credentials
from google.auth.transport.requests import Request
from google_auth_oauthlib.flow import InstalledAppFlow
from typing import Optional, Dict


class GmailClient:
    """
    Client for Gmail API.
    Handles email sending and receiving.
    """
    
    SCOPES = ['https://www.googleapis.com/auth/gmail.send']
    
    def __init__(self):
        self.credentials = self._get_credentials()
    
    def _get_credentials(self):
        """Get Gmail API credentials"""
        # Support both OAuth and service account
        cred_file = os.getenv('GMAIL_CREDENTIALS_FILE')
        if cred_file and os.path.exists(cred_file):
            return Credentials.from_service_account_file(cred_file, scopes=self.SCOPES)
        return None
    
    def send_email(
        self,
        to_email: str,
        subject: str,
        body: str,
        html: bool = False
    ) -> Dict:
        """
        Send email via Gmail.
        
        Args:
            to_email: Recipient email
            subject: Email subject
            body: Email body
            html: Whether body is HTML
        
        Returns:
            Response with message_id
        """
        try:
            from email.mime.text import MIMEText
            import base64
            
            message = MIMEText(body, 'html' if html else 'plain')
            message['to'] = to_email
            message['subject'] = subject
            
            encoded_message = base64.urlsafe_b64encode(message.as_bytes()).decode()
            
            # Placeholder for actual Gmail API call
            return {
                'message_id': 'mock_id',
                'status': 'sent'
            }
        except Exception as e:
            return {'error': str(e), 'status': 'failed'}
