alpha_loan_project/requirements.txt
"""Project requirements"""

# Core
Django==4.2.0
djangorestframework==3.14.0
django-cors-headers==4.0.0

# Database
psycopg2-binary==2.9.6

# Async & Tasks
celery==5.2.7
django-celery-beat==2.4.0
django-celery-results==2.5.0
redis==4.5.4

# AI/ML
openai==0.27.6

# APIs
requests==2.31.0
twilio==8.2.0

# Authentication
djangorestframework-simplejwt==5.2.2

# Google
google-auth-oauthlib==1.0.0
google-auth-httplib2==0.1.0

# Cache
django-redis==5.2.0

# Encryption
cryptography==40.0.2

# Environment
python-dotenv==1.0.0

# Development
black==23.3.0
flake8==6.0.0
isort==5.12.0
pytest==7.3.1
pytest-django==4.5.2
pytest-cov==4.1.0
factory-boy==3.2.1

# Documentation
drf-spectacular==0.26.1

# Monitoring
sentry-sdk==1.25.0

